Fixed a NullPointerException when fField is not set
The fField field in InlineConstantRefactoring is set in checkInitialCondition(). However since we are capturing canceled refactorings we cannot rely on that being called. Therefore, the selected affected field might not be initialized.

However, this information that is extracted is not vital. It's a simple check to see if the field has been declared as public/private. If it is private then it is marked as local change that cannot possibly affect other files.

The Refactoring history uses this information as an optimization while applying refactorings and we don't have to worry about it.