Captured rename method refactorings through the RefactoringWizard
We instrumented RenameMethodProcessor to create appropriate refactoring descriptors.

We also had to modify the way RefactoringWizard was capturing those changes; it had to be earlier, before the change was perform because making the change would invalidate some of the references i.e. fMethod.getFlags().

We also updated the txt files to show which files were modified from the original Eclipse code base.