Captured *all* RefactoringDescriptor for Rename Refactoring
We changed the mechanism for capturing the refactorings. Right now, all the classes only need to provide a createRefactoringDescriptor method and the template method in JavaRenameRefactoring will take care of the rest.

However, there is an opened issue with the way that refactoring processors are done. See Issue#96(http://github.com/vazexqi/CodingSpectator/issues/96)