Generalized the return type of the old method getCompilationUnit
In JavaWatchedRefactoring, we had a method called getCompilationUnit that returns an ICompilationUnit for use in other methods (retrieve Java project name, get code snippet, etc). We realized that we could generalize the type to its supertype i.e. ITypeRoot.

As a result of this generalization, we solve Issue #103(https://github.com/vazexqi/CodingSpectator/issues/103) so that we can also handle the type of IClassFile for InlineMethodRefactoring.