Instrumented JavaMoveProcessor
We instrumented JavaMoveProcessor to capture different move refactorings .e.g
move CompilationUnit, move files/folders. JavaMoveProcessor relies on
IMovePolicy (this is a different mechanism compared to previous refactorings).
We obtain the RefactoringDescriptor directly from IMovePolicy, cast it to a
JavaRefactoringDescriptor and call our usual methods to obtain the relevant
information.