Broke down logging of a refactoring into two steps.
First, we create a refactoring descriptor. Second, we notify the history
serializer to record the refactoring. I've exposed these two steps in Logger so
that we can easily create the descriptor before the change is created. Then, we
can serialize it whenever we're confident that the refactoring has been
performed.

This approach simplifies the logging of performed refactorings by centralizing
the code for logging the refactoring in RefactoringWizard.performFinish().
Before, we had instrumented UserInputWizardPage.performFinish(). The problem
with hacking UserInputWizardPage is that there could be multiple user input
pages in a refactorings and there could also be an error page (ErrorWizarPage).
It isn't safe to log the refactoring in the performFinish method of
UserInputWizardPage or ErrorWizarPage unless we make sure that the page is the
last page that the user sees before the refactoring is performed.
UserInputWizardPage has a method called isLastUserInputPage for this purpose.
But, I find this approach very intrusive. So, I've switched back to
instrumenting RefactoringWizard.performFinish. The problem with instrumenting
RefactoringWizard.performFinish is that the refactoring status is not complete
when we create the refactoring descriptor. Thus, we don't capture the
warning/error messages. We had this problem when we were instrumenting
UserInputWizardPage because the refactoring status is also updated during the
change creation. My suggestion is to create the refactoring descriptor before
the creation of change, then update its refactoring status after the creation of
the change. Finally, we'll serialize the descriptor after ensuring that the
refactoring has been performed.