Clean up SubmitterException and its subclasses
Made SubmitterException abstract since we never instantiate it directly.

We never use some of the constructors so I removed them. I also added a special
case for the InitializationException to test for "svn locked directory" problems.

This is done by looking for the phrase "locked; try performing 'cleanup'" – this
seems to be a standard phrase used by SVN (and SVNKit).