Refactored SVNManager into two separate classes: local and remote
Split the class SVNManager so that we have specific classes for remote
and local operations. LocalSVNManager allows operations that do not
communicate with the server. RemoteSVNManager allows both local
operations and operations that communicate with the server.
Communication with the server requires authentication information which
is not always available (or necessary).

This should prevent some possible confusion/errors when trying to invoke
remote operations on a SVNManager that does not have authentication
information.