Made the contracts between inner and outer classes of processors more explicit.
Subclasses of WatchedProcessor are inner classes of processor classes. These inner and outer classes rely on specific methods of each other. We are trying to make these dependencies explicit. So far, we have moved some abstract methods of WatchedProcessor to IWatched, IWatchedProcessor and IWatchedJavaProcessor. However, WatchedProcessor.getDescriptorID is still a code smell because it throws UnsupportedOperationException.