Added extension point to jdt.ui and functionality for replaying basic operations
The added extension point is used to notify the interested listeners that jdt.ui is about to start. This allowed to reverse the dependency: now codingtracker.recording depends on jdt.ui.

Added step by step replaying of such basic operations as creation of a new file (and the correspondig project, etc.), openning/saving/closing an editor, performing/undoing/redoing a text change.