Simplified interface of SubmitterListener
Based on our discussion today, we are going to keep the SubmitterListener interface simple for the time being. We have moved notifyPreSubmit outside of the try block and notifyPostSubmit into the finally block. notifyPostSubmit now passes a boolean to indicate whether the submission operation actually succeeded - currently no one is making use of this parameter yet.