Extracted the safe recording functionality into a separate plugin saferecorder.
Safe recording ensures that the recorder switches to a temporary file while the watched directory is uploaded to avoid writing to files that are locked during the  upload. Now this functionality is in a separate plugin and can be used both by codingtracker and codingspectator (or any other interested plugin/class). As a result, SafeRecorder is not a singleton anymore. The code that uses it should assure that only a single instance is used to log some specific information.  Besides, the plugin that contains SafeRecorder implements the extension point that notifies it when the upload is about to begin and after it is done. All instances of SafeRecorder are notified of these events (i.e. the appropriate method to switch to a temporary file and back are called).

This commit fixes the issue #149.