Captured more information for an unavailable rename unknown Java element.
Captured more context information about the location where an unavailable rename Java element is invoked from the code in `RenameJavaElementAction`. The selection and code snippet are some of the additional information that we capture.

Because `RenameJavaElementAction` may not know the exact Java element on which the refactoring was supposed to be performed, I created a new refactoring descriptor ("org.eclipse.jdt.ui.rename.unknown.java.element") to use as the ID of such unavailable refactorings.

Also, I added a test that verifies what I implement in this commit by trying to rename a Java keyword.

This commit belongs to issue #163.