Added back the missing "selection" attribute.
Renamed the "selection" attribute in `RefactoringDescriptor` to "selection-text".

Renamed the "selection-offset" attribute in `RefactoringDescriptor` to "selection-in-code-snippet".

Restored the statement that set the attribute "selection" in the `populateRefactoringSpecificFields` methods of `ExtractMethodRefactoring`, `InlineConstantRefactoring`, `InlineTempRefactoring` and `InlineMethodRefactoring`.

Made the automated UI test verify the descriptors that CodingSpectator captures for valid/invalid canceled/performed invocations of the extract method refactoring.

This commit belongs to issues #168 and #138.