Cleaned up the tests for extract method refactoring.
Changed ExtractMethodRefactoring.getSimpleRefactoringDescriptor to take the union of the arguments of the original descriptor with that of WatchedJavaRefactoring.populateInstrumentationData.

Note that WatchedJavaRefactoring.populateRefactoringSpecificFields gets called twice now. Even though it's redundant to call this method twice, it does no harm.

Parameterized the package name in the expected values of attributes.

Cleaned both the CodingSpectator and Eclipse refactoring log files in the tests for performed extract method refactorings.

This commit belongs to issue #138.