Listened to the errors reported to Platform.
Changed the log object CodingSpectator listens to in order to capture all the errors that show up in the "Error Log View". Currently, we believe that CodingSpectator captures all errors except those that get raised before we begin listening to Platform in the early startup event.

Added a plugin called edu.illinois.codingspectator.errorlog.trigger to generate different kinds of log events. We use this plugin for testing the capturing of error log in CodingSpectator.

This commit belongs to issue #143.