Fixed the failures pf RenameType tests.
Updated the status using the latest information about the refactoring being performed. The updated status helped prevent the assertion failure because I used the updated status to decide whether to check the changes performed on the workspace as a result of the refactoring.

I've tried to keep my changes faithful to the original refactoring tests in JDT.

All the tests in org.eclipse.jdt.ui.tests.refactoring.AllRefactoringTests pass as a result of this commit.

This commit belongs to issue #164.