Set focus to password field if the username is "remembered"
Previously the code always sets focus to the username field. Howeer, this becomes problematic if the username is "remembered" i.e. the user has already entered it once before. So now we check to see if the username is set already or not and choose to focus on the password field instead.

This commit belongs to issue #172.