Simplified the verification of captured refactoring descriptors.
We used to verify every attribute of captured refactoring descriptors in assert statement individually. This approach was very time consuming and hard to maintain. So, we switched to a more automatic mechanism for checking the contents of captured refactoring descriptors. Basically, we store the expected refactoring descriptor as auxiliary files in the test project. Then, we parse both the expected and actual refactoring descriptors and compare them. Of course, we need to make sure that the tests are not sensitive to the differences in timestamps. Therefore, we set the timestamp in the navigation-history attributes of the expected and actual refactoring descriptors to -1. And, we just check that the timestamps captured in the refactoring descriptors are positive numbers.

This commit belongs to issue #138.