Made SafeRecorder more resilient and generic.
Now SafeRecorder ensures that the record file exists before every recording such that in case a file and/or its parent folders are deleted somehow, they will be recreated upon the following recording. Also, error messages now mention the actual problematic file (SafeRecorder could be used for different tasks now, so it was not enough just to say "main" file or "temporary" file).