Shortened the length of paths in edu.illinois.codingspectator.ui.tests.
Windows has a limit of 255 characters on the length of a path. The following was the longest path in CodingSpectator and had 319 characters.

/edu.illinois.codingspectator.ui.tests/expected-refactoring-descriptors/extract-interface/ValidPreviewedAndPerformedExtractInterfaceParentClassTest/performed/TestProject_class edu.illinois.codingspectator.ui.tests.extractinterface.ValidPreviewedAndPerformedExtractInterfaceParentClassTest/2011/3/13/refactorings.history

This commit shortens the above path to the following path of length 242.

/edu.illinois.codingspectator.ui.tests/expected-descriptors/extract-interface/ValidPreviewedAndPerformedExtractInterfaceParentClassTest/performed/Project_ValidPreviewedAndPerformedExtractInterfaceParentClassTest/2011/3/13/refactorings.history

Renamed the directory "expected-refactoring-descriptors" to "expected-descriptors". And, changed the naming of the generated test projects from "TestProject_<fully qualified name of the test class>" to "Project_<simple name of the test class>".

This commit breaks all of the tests except CancelledExtractConstantTest because the expected descriptors of the other tests need to get updated to the new test project names.

This commit belongs to issue #180.