Persisted the preference store of the automatic update plug-in.
Removed minimum version requirements from the dependencies of the monitor.ui plug-in to fix a dependency error during the installation of CodingSpectator.

Saved the preference store of the automatic update plug-in after enabling the update notifications. Without saving the preference store, the update notifications will get disabled when Eclipse is restarted.

This commit belongs to issue #173.