Fixed a compilation error with T25 and T26
In doExecuteRefactoring, T25 and T26 throw Exception. However, the
method signature for doExecuteRefactoring says that its throws
CoreException; so the types are incompatible.

I checked the method that was throwing the Exception and it
was createANewJavaClass in CodingSpectatorBot. However, upon closer
inspection, I discovered that it does not need to throw any exception.

So I removed the throws Exception from the method signature.

This fixes the compilation error that I had.