Initial attempt at capturing post refactoring problems
We have operated under the assumption that the "Build Automatically" option is turned on to keep things simple for the time being. Once we have finished this part we can proceed with trying to capture org.eclipse.jface.text.source.Annotation from the Java Editor.

The TODO's how what else needs to be done.

This commit belongs to issue #191.