Report problem changes between pre and post refactoring
Significant changes:

* Added hashCode() and equals() so we can compare IProblems
* Stored a previous/current problems for diffing after refactoring
* Called CompilationUnit#becomeWorkingCopy to prevent the reconciler
operation from complaining that a file does not exist.

Problems:

* IProblem keeps track of its location of the problem in the code. After
a refactoring is performed, some of the problems "shift" around to
different lines; this means that our diffing method does not work to
identify changed problems.

Proposed solution:

* Do a symmetric set diff both ways A\B and B\A so we can keep track of
problems introduced and removed by a refactoring.
* We still have to use manual inspection to see if two problems "match"
before and after the refactoring.

This commit belongs to issue #191.