Files are read and written using the appropriate encoding.
The workspace files are read using the corresponding encoding (or the default, if none can be obtained). The files that are produced by the tool (e.g. operation sequence) are read and written using UTF-8, which should be used for the replaying as well.

This commit addresses item (1) of issue #177.