Captured more precisely selection info at invocation time
We capture and store the selection information at the time of refactoring invocation (the Action classes). Then during the refactoring logging processwe access the information stored to save it into the descriptor.

Right now, we have only done this for Pull Up Action.

CAUTION: Test cases have not been fixed to reflect changes yet.

This commit belongs to issue #170.