Captured invocation information through Structured View
We now can capture relevant information when a refactoring is invoked through the structured view e.g Outline View and Package Explorer. We added a new attribute to the refactoring descriptor to indicate that the refactoring was invoked throgh this mechanism.

CAUTION: Tests are still broken.

This commit belongs to issue #170.