Cleared RefactoringGlobalStore after serializing the refactoring in Logger.
Moved the call to RefactoringGlobalStore.clearData to Logger. This frees us from clearing the global store whenever we use its contents and prevents us from clearing the global store before we are completely done with it. By clearing the global store after the refactoring is serialized by Logger, we are sure that we don't clear the RefactoringGlobalStore before we are done with it.  Another benefit of clearing the RefactoringGlobalStore in Logger is that it lets us clear the RefactoringGlobalStore in a single place.

However, if a refactoring fails such that CodingSpectator doesn't serialize it, the RefactoringGlobalStore will go to an inconsistent state. One way to prevent the RefactoringGlobalStore from becoming inconsistent is to get a new instance of the RefactoringGlobalStore at the beginning of the run methods of the action classes of all refactorings that might use the RefactoringGlobalStore.

This commit belongs to issue #163.