Initialized RefactoringGlobalStore in the run methods of refactoring Actions.
By initializing the RefactoringGlobalStore, we make sure that CodingSpectator does use the data stored in RefactoringGlobalStore by previous refactorings. Besides, by using the selection information stored in the RefactoringGlobalStore, we gain more precision as sometimes the offsets stored in WatchedJavaRefactoring do not precisely represent the selections that the user has made.

This commit belongs to issue #163.