Extracted the code snippet from the unchanged version of the source code.
The refactoring may change the buffer that is used to compute the code snippet and the relative offsets. Therefore, I changed the code so that it uses the contents of the file rather than the ITypeRoot object to compute the code snippet and the offset of the selection relative to the code snippet.

This commit belongs to issues #163 and #207.