Created well-formed problems XML tag
Previously when we serialize the XML problem tag we were only using the printTag method. This did not append a corresponding end tag so it will throw an exception when we tried parsing it with a XMLParser.

Now we use startTag(...) and endTag(...) to wrap the problem element properly.

This commit belongs to issue #191.