Constructed a parser for reading through the refactoring log
The refactoring problems log contains appended problem-changes. However, because we are always appending new problem-changes, we do not include the necesarily end XML tag. This causes problems for XML parsers. Therefore we make use of CompositeReader and wrap our refactoring problems log with faux start and end XML tags when we read.

This commit belongs to issue #191.