Tweaked the replaying infrastructure to support 2 modes of operations replaying.
User operations can be replayed in 2 modes. In TEST_MODE user operations are replayed in the same way, in which they happened, i.e. refactorings are replayed, undone and redone text edits are replayed through the document udo manager, etc. In general, this mode simulates the previous recorder/replayer behavior. This mode has to be set as the corresponding environment variable, and it is used for running tests and replaying of the old operation sequences.

In normal mode (this is the mode when TEST_MODE is NOT set) replayer reproduces effects of refactorings rather than trying to replay them, also undone and redone operations are replayed as performed operations. This is a more robust mode and it is designed for the new version of the tool, which records low level atomic changes produced by refactorings.

This commit belongs to issue # 203.