Discarded working copies that are created by CodingSpectator.
The current mechanism for computing the problems of a compilation unit requires that the compilation unit be a working copy. So, CodingSpectator makes the affect compilation units working copies. But, this has undesirable side effects. Therefore, we changed CodingSpectator to discard the working copies that it creates after computing the problems.

An affected compilation unit might not exist when CodingSpectator tries to compute the problems. So, we added the checks to avoid using compilation units that don't exist for computing the problems.

This solution is not ideal because it has side effects that might change the Eclipse behavior in unexpected ways. So, we have to investigate safer ways to compute the problems of compilation units.

This commit belongs to issue #210.