Updated CodingSpectator's code to use newer InstanceScope API
Previoulsy in  EPP, we were manipulating the preference store using the older API which has deprecated. This commit updates the places that CodingSpectator has touched to use the newer InstanceScope API. There are still places in EPP that use the older API and we have left those as is to localize our changes only to the functions that we touched.

This commit belongs to issue #214.