Made the timestamps in refactoring-problems.log consistent with those of Eclipse
Added an attribute "refactoring-timestamp" to the node "problem-changes" of "refactoring-problems.log". The timestamps stored in this new attribute matches the timestamps that Eclipse uses in the refactoring descriptors that it generates for performed refactorings. Using the same timestamps makes it very easy to relate the compilation errors reported in "refactoring-problems.log" to the refactoring descriptors captured by Eclipse.

It is nontrivial to match the timestamps of the descriptors that CodingSpectator generates for performed refactorings with the ones that Eclipse uses in its corresponding refactoring descriptors. So, I didn't match up these two kinds of timestamps. But, I improved the timestamp generation for the descriptors created by CodingSpectator in two ways. I removed the generation of timestamp in the constructor of RefactoringDescriptor and moved it to the serialization of the descriptor in the Logger class. Also, I made the cloning mechanism of refactoring descriptors preserve the timestamps.

This commit belongs to issue #120.