Added the remaining logs files for expected refactoring problems.
I temporarily changed RefactoringTest to copy the generated refactoring log file to the corresponding expected file. It took about an hour for the tests to complete. Having generated the expeccted log files, I reverted the changes to RefactoringTest and added the generated refactoring problems for regression testing.

This commit belongs to issue #194.