Captured the code snippet information immediately after the selection.
Computed the code snippet information right after making a selection for a refactoring and stored the result in `RefactoringGlobalStore`. Used the buffer of the `ITypeRoot` object rather than the underlying `IFile` object to compute the code snippet information because the contents of the `IFile` object may be out of date. Since CodingSpectator is now using the `IBuffer` to compute the code snippet information, it's important to compute and save the code snippet information right after the selection is made and before the refactoring is invoked. If we delay the computation of code snippet information, CodingSpectator may face problems in the extraction of code snippet information because the `IBuffer` may change due operations such as preview (See issue #207).

This commit belongs to issue #232.