Ensured that CodingTracker does not listen two documents of the same file.
In some scenarios for code assist (e.g. creating a new inner file) Eclipse creates org.eclipse.jdt.internal.core.DocumentAdapter for a buffer that is org.eclipse.jdt.internal.ui.javaeditor.DocumentAdapter. As a result, CodingTracker is registered for both DocumentAdapters and records duplicated code changes for the same file. To avoid this problem, CodingTracker explicitly checks the class of the buffer and registers a listener to its document only if this buffer is an instance of Buffer class.

This commit fixes the second part of issue #247.