Captured UDC data on Eclipse SDK.
Eclipse SDK doesn't come with UDC. So, CodingSpectator has to install UDC in order to be able to capture the UDC data. This commit includes the UDC feature in the CodingSpectator feature.

Note that the feature org.eclipse.epp.usagedata.feature has to be present and open in the workspace in which the update site is built.

This commit belongs to issue #214.