Replaced some syncExec's by asyncExec's.
I reviewed all jobs that are run in the UI thread synchronously. I replaced syncExec by asyncExec wherever possible and added a comment explaining the reason for using syncExec in other cases.

Added a tool tip to the check box on the "Display" preference page that controls the appearance of the logo in the status line.

This commit belongs to issue #281.