SafeRecorder now stores files outside of the watched directory.
Record files are stored outside of the watched directory. They are copied into the watched directory right before an upload is about to happen. Files that are copied successfully are deleted from the storage location. This is needed to avoid copying temporary files over and over again as well as to avoid copying too big main record files (i.e. currently, the main record file would contain only the delta that has accumulated since the last upload).

This commit belongs to issue #257.