Captured the infer type arguments refactorings for Helios.
Test infertypearguments.T03 changes the configuration of the input page of the refactoring. So, I made the test invoke the refactoring twice to revert the changes in the second invocation. Otherwise, the configuration change would have got propagated to future invocations of the refactoring. Because this test logs two refactorings, I extended the checkers to handle log files containing several refactorings.

This commit belongs to issue #272.