CodingTracker tracks changes to conflict editors even not knowing edited file.
Now, even if the underlying Java file can not be detected when a conflict editor is opened, CodingTracker still records operations belonging to this conflict editor. To replay such operations, a postprocessing step is needed, which will detect the edited Java file from other related operations and update the conflcit editor open operation correspondingly.

This commit belongs to issue #286.