Fixed a bug in the CSV generator and added more types of events to the reports.
Fixed a bug that excluded UNDONE and REDONE refactoring events of CodingTracker from the CSV files.

Added refactoring events from the old and new formats of the CodingTracker's logs.

Added a column to the CSV reports, called "recorder", to indicate the source of the data (Eclipse, CodingSpectator and CodingTracker).

This commit belongs to issue #293.