Added a way to extract location information from RefactoringDescriptorProxy
Unfortunately RefactoringHistoryManager is mostly a static-methods class and as such I have used this turn on/off mechanism to enable this functionality.

The downside of this is that you need to remember to turn it off.  Note that leaving it on is not a problem; it just might do additional work compared to the original Eclipse version.

RefaactoringHistoryManager.setPopulateProxiesWithRefactoringDescriptorFolder();
// Turn on storing and extracting of file path (timestamp) in
// RefactoringDescriptorProxy
....
RefactoringHistoryManager.unsetPopulateProxiesWithRefactoringDescriptorFolder();
// Turn off storing and extracting of file path (timestamp) in
// RefactoringDescriptorProxy

This still needs further testing. I have not tested in on the full
dataset yet.

This commit belongs to issue #298.