Pathched LTK Core in Helios
The previous commits only affect Indigo. Technically we don't need to patch the Helios files since what we are doing is only for data analysis. And for data analysis we can just run that using Indigo.

However, for consistency and uniformity across both Helios and Indigo, it's better to make those changes e.g. when we are trying to diagnose a bug at least we have similar behavior across both Helios and Indigo.

This commit belongs to issue #298.