Added test case for locating RefactoringDescriptors
This is a copy of test04 with the name of the dates changed. This does not affect the parsing because we now rely on the new mechanism for getting the location of the RefactoringDescriptor from the information stored in the RefactoringDescriptorProxy during the initial scan of the directories.

This commit belongs to issue #298.