Externally changed files are deleted in order avoid problems with commits.
An external change operation may be followed by a commit operation, which compares the committed file content against the file content in the workspace, if the file exists. Therefore, externally changed files should be always deleted (except in testing mode) to ensure that the following commit operations do not fail.