A newly created compilation unit is saved if it is opened in an editor.
If a compilation unit is created after an external change to the corresponding file, it might be already opened in an editor. As a result, the editor will become dirty and the new content may be lost if the user did not save the editor explicitly. Therefore, any newly created compilation unit is saved upon creation if there is an open editor displaying the corresponding underlying file.