Projects are now deleted both from the workspace and from the disk.
Regardless of the actual options, projects are erased completely. This is needed in order to avoid situations when a project is deleted from the workspace only (i.e. preserving its content on the disk), then the files on the disk are deleted manually (which is missed by CodingTracker), and then, there is some move operation that tries to move something into the previously deleted project and fails, because the destination project's content still exists during the replay.