Changed the output folder name to some unusual name to avoid name collisions.
Before, the output folder name was 'bin'. The output folder is created every time a new project is created. As a result, some operations that copy or move 'bin' folder failed because the destination folder already existed. To avoid this name conflict, CodingTracker now uses an unusual name for the output folder.