Change CodingTrackerPostprocessor to use env variable
Instead of coding the path of the directory to analyze, it might be
better to obtain that through an environment variable. This avoid the
issue where we might accidentally change the directory and then commit
that change.

Of course this would require that you set the environment variable in
the Environment Tab in the JUnit Plug-in Test configuration.