Added a table to combine the change size and configuration time of refactorings.
Adde a table called CHANGE_CONFIGURATION_COUNT to report both the size and configurtaion time of refactoring. However, column CONFIGURATION_TIME_IN_MILLI_SEC of the resulting table is always "[null]". I've used a subquery to populate the table. There seems to be a problem in the subquery, but, I haven't been able to debug it.

This commit belongs to issue #322.