Split RAW MODE input into smaller chunks
I suspected that the input we had in RAW MODE for the creation of functions was beginning to get too big and might have caused some buffer/sending issues when we finally commit. Therefore, I split the function creation section into smaller sections.

Doing so seems to fix the "object not found or user lack privilege" errors that I was seeing this afternoon. I verified that this works on a clean database.

This commit belongs to issue #288.