Add username to JavaRefactoringChangeSizeConfiguration CSV
In addition to the current parameters, this commit adds the username to the generated files for correlation analysis. Adding the username allows us to check if one particular invocation of a refactoring tool was dominated by a particular user. This statistics is useful to know when we run the correlation analysis to see if we have statistically significant data for the tool usage.

This commit belongs to issue #288.