Count CodingSpectator refactorings that were undone
Because of a weird bug with hsqldb and correlated subqueries (nested) we have to generate the RAW SQL to determine which CodingSpectator captured refactorings were actually undone (via a join with CodingTracker captured refactorings that are close in execution time).

The RAW SQL statements are generated using Ruby using string substitutions. If we need something more advanced we can use ERB.

This commit belongs to issue #288.