Used the SVN update operation to update the local revision numbers.
The SVN commit operation does not automatically increase the revision number of the local repository. So, we have to update the local revision numbers explicitly in order to make the detection of outdated working copies easy. I used the SVN update operation to update the local revision numbers instead of the SVN checkout operation. It is safe to do an SVN update immediately after an SVN commit because there is a rare chance of introducing a conflict in such a short time interval.

This commit belongs to issue #311.