Updated the path to actual refactoring logs in unit tests.
Because CodingSpectator moved the refactoring histories outside the watched folder, the tests couldn't find the actual refactoring logs because they expected the logs to be in the watched folder. So, I triggered the notifications of SubmitterListener just to gather them all in the watched folder.

This commit belongs to issue #287.