Added a graph and fixed a bug in the SQL script.
I realized that "SELECT COUNT(DISTINCT(COLUMN))" does not ignore NULL values. As a result, the total number of programmers who have used the refactoring tool is 195,105 not 780,427. I updated the SQL script to report the number of refactoring users correctly.

Renamed table "UDC_REFACTORINGS" to "UDC_EVENTS".

Added a graph for the distribution of the number programmers over the number of refactorings they have invoked.