Record CoherentTextChanges instead of original ones in AST inference.
Instead of the original text changes, the AST operations inferencer records the corresponding CoherentTextChanges. As a result, the inferred AST operations are better aligned with the causing textual changes. This approach facilitates inference of manual refactorings, in particular, handling of boxed (aka linked) edits. Also, it reduces the amount of operations to be replayed and simplifies the subsequent AST inference (e.g., while inferring manual refactorings), which both lead to better performance.