Added support for corrective refactoring properties.
Generalized the inference infrastructure to support corrective refactoring properties. Such properties do not serve as parts of the inferred refactorings, but rather correct values of already accumulated refactoring properties.

Currently, implemented a corrective property that updates simple names. This allows to correctly infer refactorings when, for example, the added element's name (e.g., the method name in an extracted method declaration) is changed after the element is added.