Added more checks to esnure that a new covering node is well formed.
If a new covering node is a Block that represents the body of a method, then even if the body is malformed, the Block will not be marked as malformed. Instead, the method's declaration will be marked as malformed. Therefore, an additional check is needed to ensure that a new covering node is indeed well formed.