All AST operations related to an inferred ref. are marked with its ID.
Expanded the infrastructure to track AST operations that are related to AST operations that form properties out of which inferred refactoring are constructed. Such related AST operations may precede or succeed the property-forming operation, be in the same batch of AST operations or not, etc.

Also, made many other improvements and fixes, e.g., now Rename Class refactoring looks not only for changed references but also for changed constructor names.