Considered old/new values of a changed node as being deleted/added.
To infer move operations, the old value of a changed node is considered to be deleted, while the new value is considered to be added. Also, introduced an additional ID to the move operation: moveID is used for the moved changed nodes whose values are added, while deletingChangeMoveID is used for the moved changed nodes whose values are deleted. These IDs are used to generated the corresponding "move" properties during refactoring inference.