Represented itemset frequencies as lists rather than scalar values.
This allows to decide whether adding a new item to an itemset reduces the itemset's "gluing power" or not. The "gluing power" of an itemset is its potential capacity to absorb additional items, while keeping the overall frequency at the maximum. The algorithm was updated such that all optimizations do not decrease the "gluing power" of the optimized itemsets.